Fixed issues with putItem
using put would place the next item in the inventory down instead of the intended item. This is an issue that was caused by vector.erase() deleting the original Item object in the player's inventory.  Then, as the vector resizes, the next item in the vector takes the address of the deleted item, which is still pointed to by itemPtr. So the item that is placed in the room is the item that now assumes this position